Переробила структуру проекта
