+++ controllers/contacts/addContact.js
@@ -1,0 +1,8 @@
+const { Contact } = require("../../models");
+
+const add = async (req, res) => {
+  const newContact = await Contact.create(req.body);
+  res.status(201).json(newContact);
+};
+
+module.exports = add;
