+++ controllers/index.js
@@ -1,0 +1,5 @@
+const ctrl = require("./contacts");
+
+module.exports = {
+  ctrl,
+};
