+++ routes/index.js
@@ -1,0 +1,3 @@
+const contactsRouter = require("./api/index");
+
+module.exports = { contactsRouter };
